# affiliate/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Paparazzi Affiliate Commerce Platform — v1 API
  version: 1.0.0
  description: |
    Public v1 API for the Paparazzi Affiliate Commerce Platform (looks, offers,
    links, integrations, earnings, payouts, programmes, publishers, disputes,
    contracts, suspense).

    This spec is generated from the real route handlers in
    `packages/api/src/routes/*.ts` — it documents only endpoints the code
    registers. There is no `GET /v1/links` (links are minted, then resolved
    by the redirect service at `GET /r/{token}`).

    ## Authentication (temporary stub)

    All `/v1/*` endpoints require a Bearer JWT (`Authorization: Bearer <token>`).
    Expected claims: `sub`, `org_id`, `role`. **The JWT verification in
    `src/middleware.ts` (`requireAuth`) is a temporary stub: it trusts the
    token's claims verbatim and must be replaced by a production identity
    provider** with membership validation (`memberships` table), issuer /
    audience / expiry policy. `GET /healthz` is the only unauthenticated route.

    ## Envelopes

    - Success: `{ "data": <payload>, "request_id": "<uuid>" }` (`ok()` helper).
    - Error: `{ "error": { "code": "<ERROR_CODE>", "message": "..." }, "request_id": "<uuid>" }`.
    - Every response carries an `X-Request-Id` header echoing `request_id`.
    - Error `code` is always one of the `ErrorCode` enum values
      (`packages/shared/src/errors.ts`); clients should branch on `code`,
      never on message text.

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        TEMPORARY STUB — `requireAuth` in `src/middleware.ts` verifies the JWT
        signature against `JWT_SECRET` and trusts the `sub`, `org_id`, `role`
        claims verbatim. Replace with a production IdP (membership validation,
        issuer/audience/expiry policy) before pilot.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 200
      description: |
        Optional idempotency key. First response per (key, org) is stored and
        replayed with `X-Idempotent-Replay: true`; 5xx is never stored.
...
  headers:
    X-Request-Id:
      description: Echoes the envelope `request_id` for log correlation.
      schema:
        type: string
        format: uuid
    X-Idempotent-Replay:
      description: Present (`true`) when the body is a replay of a stored idempotent response.
      schema:
        type: string
        enum: ['true']
  schemas:
    ErrorCode:
      type: string
      enum:
        - PROGRAMME_NOT_APPROVED
        - OFFER_STALE
        - PROPERTY_FORBIDDEN
        - PUBLISHER_NOT_ACTIVE
        - PROPERTY_NOT_OWNER_OPERATED
        - NOT_FOUND
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - CONFLICT
        - RATE_LIMITED
        - UPSTREAM_UNAVAILABLE
        - LEDGER_IMBALANCE
        - TRANSFER_STATUS_UNKNOWN
        - INTERNAL
      description: Canonical codes from `packages/shared/src/errors.ts`.
    ErrorEnvelope:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
        request_id:
          type: string
          format: uuid
    SuccessEnvelope:
      type: object
      required: [data, request_id]
      properties:
        data: {}
        request_id:
          type: string
          format: uuid
      description: Shape of every 2xx body; `data` is operation-specific.
    Uuid:
      type: string
      format: uuid
...
    InternalConversionStatus:
      type: string
      enum: [received, pending, approved, declined]
    LedgerDisposition:
      type: string
      enum: [posted, skipped]
      description: |
        `posted` — ledger entries were written; `skipped` — conversion left
        unposted (unattributed click or no approved contract; never guessed).
    # ---- domain models ----
...
    RowError:
      type: object
      required: [row, reason]
      properties:
        row: { type: integer, description: 1-based data-row number; 0 = file-level error. }
        reason: { type: string }
  responses:
    BadRequest:
      description: '400 — request failed validation (VALIDATION_ERROR).'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: '401 — missing/invalid Bearer token or claims (UNAUTHORIZED).'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: '403 — role not allowed (FORBIDDEN), or a link-guard rejection (PROPERTY_FORBIDDEN, PROGRAMME_NOT_APPROVED, PUBLISHER_NOT_ACTIVE, PROPERTY_NOT_OWNER_OPERATED).'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: '404 — no route, unknown connector, or resource not in this org (NOT_FOUND).'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    InternalError:
      description: '500 — unexpected failure (INTERNAL); message not leaked.'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
paths:
...
  /v1/integrations/csv/uploads:
    post:
      tags: [Integrations]
      summary: Upload merchant settlement CSV
      description: |
        File-based merchant settlement import (brief §8). The CSV is POSTed
        inline as a plain string — no multipart (2 MB cap; a deliberate
        sandbox choice). The uploader names `programme_id`, which must exist
        in the org and be `active` (`404` / `422` otherwise). An Amazon.in
        Associates programme, or a `provider_account_id` that is an Amazon
        account's, is `422 VALIDATION_ERROR`: those rows come only through
        `POST /v1/integrations/amazon-associates/reports`.

        Every row goes through the SAME conversion pipeline as the provider
        webhook: idempotency on `(provider_account_id, source_transaction_id,
        line_id)`, `provider_revision` ordering, unknown `returned_click_ref` →
        suspense (`click_id` NULL), ledger posting on approved, same outbox events.

        **All-or-nothing validation**: every row is validated before anything
        is ingested. One bad row → `422 VALIDATION_ERROR` with per-row
        `errors: [{row, reason}]` and zero inserts. `row` is the 1-based
        data-row number (header not counted; blank lines skipped; `0` = file-level).

        Columns (header required, case-insensitive, extra ignored, RFC-4180,
        UTF-8, LF/CRLF): `source_transaction_id` (req), `line_id`,
        `returned_click_ref`, `currency` (req, 3-letter), `eligible_value_minor`
        (req, integer ≥ 0 — decimals rejected), `commission_minor` (req,
        integer ≥ 0), `provider_status` (req: `approved|pending|declined|reversed`,
        case-insensitive; `reversed` behaves as `declined` post-approval),
        `provider_revision` (int ≥ 0, default 0), `occurred_at` (req, ISO 8601).

        Roles: `network_admin`, `editor`.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider_account_id, programme_id, filename, csv_text]
              properties:
                provider_account_id: { type: string, minLength: 1, maxLength: 200 }
                programme_id: { $ref: '#/components/schemas/Uuid', description: The uploader names the programme (contrast the webhook path, which resolves it). }
                filename: { type: string, minLength: 1, maxLength: 255 }
                csv_text: { type: string, minLength: 1, description: Inline CSV text; 2 MB max. }
            examples:
              upload:
                value:
                  provider_account_id: direct-merchant-009
                  programme_id: 33333333-3333-4333-8333-333333333333
                  filename: settlement-2026-09-20.csv
                  csv_text: "source_transaction_id,currency,eligible_value_minor,commission_minor,provider_status,occurred_at,returned_click_ref\ntxn-0001,INR,249900,16000,approved,2026-09-20T10:00:00.000Z,clk_9f2c1ab4\n"
      responses:
        '202':
          description: File ingested; per-row outcomes returned.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
            X-Idempotent-Replay: { $ref: '#/components/headers/X-Idempotent-Replay' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        type: object
                        required: [filename, provider_account_id, programme_id, rows_received, created, deduped, status_changed, auto_reversed, results]
                        properties:
                          filename: { type: string }
                          provider_account_id: { type: string }
                          programme_id: { $ref: '#/components/schemas/Uuid' }
                          rows_received: { type: integer }
                          created: { type: integer }
                          deduped: { type: integer }
                          status_changed: { type: integer }
                          auto_reversed: { type: integer }
                          results:
                            type: array
                            items:
                              type: object
                              required: [row, outcome, conversion_id, status]
                              properties:
                                row: { type: integer }
                                outcome: { type: string, enum: [created, deduped, status_changed, auto_reversed] }
                                conversion_id: { $ref: '#/components/schemas/Uuid' }
                                status: { $ref: '#/components/schemas/InternalConversionStatus' }
                                ledger: { $ref: '#/components/schemas/LedgerDisposition' }
                                auto_reversed: { type: boolean }
                                adjustment_id: { $ref: '#/components/schemas/Uuid', nullable: true }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: |
            `VALIDATION_ERROR` with top-level `errors: [{row, reason}]`
            (the shared `ErrorCode` union has no CSV-specific code, so the
            per-row detail rides in `errors`). Nothing was ingested.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorEnvelope'
                  - type: object
                    required: [errors]
                    properties:
                      errors:
                        type: array
                        items: { $ref: '#/components/schemas/RowError' }
              examples:
                row-errors:
                  value:
                    error: { code: VALIDATION_ERROR, message: '2 row(s) failed validation; nothing was ingested' }
                    request_id: 9f2c1ab4-e5d6-4f7a-8b9c-0d1e2f3a4b5c
                    errors:
                      - { row: 2, reason: "commission_minor must be an integer in minor units (no decimals); got '12.50'" }
                      - { row: 3, reason: "provider_status must be one of approved|pending|declined|reversed; got 'bogus'" }
        '500': { $ref: '#/components/responses/InternalError' }
...